Split noise refresh decrypt and merge circuits
