+++ src/noise_refresh/mod.rs
@@ -1,2 +1,3 @@
-pub mod circuit_format;
+pub mod circuit_decrypt;
+pub mod circuit_merge;
 pub mod circuit_prg;
